+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Improvements
 - Make use of the new consuming variants in InputHandler and Keybind to consume UiControls inputs
 - Allow Tooltip to manage more than one paragraph and make it easier to add new lines
 - Allow adding dropdown elements at a specified index
+- Turned Tooltip paragraph styling into style properties
 
 Fixes
 - Fixed auto-nav tooltip displaying on the selected element even when not in auto-nav mode